Base functionality VotingComponent
Notes:  Used AtomicInteger because Integer() is annoyingly immutable.
All example XML files he gave us work. 701a.xml crashes because there isn't a candidate15, so just change the candidate value to an existing candidate to test.
Only thing left to do is dealing with the passcodes to initializeTallyTable/getRankedReport. I didn't do it because I'm not sure if we're just supposed to compare the passcode to a hardcoded String or w/e